requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
web3==6.4.0
requests==2.31.0
